+++ src/nativeMain/kotlin/process/InitProcess.kt
@@ -120,7 +120,6 @@ private fun initProcessInternal(
                                 "spec.process.terminal=true but --console-socket was not provided",
                             )
                             _exit(1)
-                            @Suppress("UNREACHABLE_CODE")
                             return@memScoped
                         }
                 consoleSocketFd = connectConsoleSocket(consoleSocketPath)
@@ -284,7 +283,6 @@ private fun initProcessInternal(
                         Logger.error("failed to allocate pseudo-terminal from container devpts")
                         if (consoleSocketFd >= 0) close(consoleSocketFd)
                         _exit(1)
-                        @Suppress("UNREACHABLE_CODE")
                         return@memScoped
                     }